BOAC-5024, move Google Analytics utils to a dedicated ga.ts file
