+++ package.json
@@ -10,13 +10,14 @@
   "dependencies": {
     "@ckeditor/ckeditor5-build-classic": "41.0.0",
     "@ckeditor/ckeditor5-vue2": "3.0.1",
-    "@fortawesome/fontawesome-svg-core": "6.4.2",
-    "@fortawesome/free-regular-svg-icons": "6.4.2",
-    "@fortawesome/free-solid-svg-icons": "6.4.2",
+    "@fortawesome/fontawesome-svg-core": "6.5.1",
+    "@fortawesome/free-regular-svg-icons": "6.5.1",
+    "@fortawesome/free-solid-svg-icons": "6.5.1",
     "@fortawesome/vue-fontawesome": "2.0.10",
     "@trevoreyre/autocomplete-vue": "2.4.1",
+    "@types/gtag.js": "^0.0.18",
     "@vue-a11y/announcer": "2.1.0",
-    "axios": "1.6.5",
+    "axios": "1.6.7",
     "bootstrap-vue": "npm:@neverendingsupport/bootstrap-vue2@2.23.2",
     "d3": "5.16.0",
     "highcharts": "9.3.1",
@@ -41,19 +42,18 @@
   "devDependencies": {
     "@babel/core": "^7.23.3",
     "@jridgewell/source-map": "^0.3.5",
-    "@types/gtag.js": "^0.0.18",
     "@vue/cli": "^5.0.8",
     "@vue/cli-plugin-babel": "^5.0.8",
     "@vue/cli-plugin-eslint": "^5.0.8",
     "@vue/cli-plugin-typescript": "^5.0.8",
     "@vue/cli-service": "^5.0.8",
     "@vue/eslint-config-typescript": "^12.0.0",
-    "eslint": "^8.55.0",
-    "eslint-plugin-vue": "^9.19.2",
-    "eslint-plugin-vue-scoped-css": "^2.5.1",
+    "eslint": "^8.56.0",
+    "eslint-plugin-vue": "^9.20.1",
+    "eslint-plugin-vue-scoped-css": "^2.7.2",
     "npm-force-resolutions": "^0.0.10",
     "typescript": "^5.2.2",
-    "vue-eslint-parser": "^9.3.2",
+    "vue-eslint-parser": "^9.4.2",
     "vue-template-compiler": "npm:@neverendingsupport/vue2@2.7.17-template-compiler",
     "vue-tsc": "^1.8.25"
   },
