+++ src/berkeley.ts
@@ -1,6 +1,5 @@
-import _, {map} from 'lodash'
+import _ from 'lodash'
 import store from './store'
-import Vue from 'vue'
 import moment from 'moment'
 
 export function describeCuratedGroupDomain(domain, capitalize) {
@@ -103,7 +102,7 @@ export function getCsvExportColumns(domain) {
 }
 
 export function getCsvExportColumnsSelected(domain) {
-  return domain === 'default' ? ['first_name', 'last_name', 'sid', 'email', 'phone'] : map(getAdmitCsvExportColumns(), 'value')
+  return domain === 'default' ? ['first_name', 'last_name', 'sid', 'email', 'phone'] : _.map(getAdmitCsvExportColumns(), 'value')
 }
 
 export function getDefaultCsvExportColumns() {
@@ -197,47 +196,6 @@ export function hasMatrixPlottableProperty(obj, prop) {
     return !!getMatrixPlottableProperty(obj, prop)
   }
   return _.isFinite(getMatrixPlottableProperty(obj, prop))
-}
-
-export function initGoogleAnalytics() {
-  return new Promise<void>(resolve => {
-    const googleAnalyticsId = process.env.VUE_APP_GOOGLE_ANALYTICS_ID
-    const user = store.getters['context/currentUser']
-    const uid = user.uid
-    if (googleAnalyticsId) {
-      window.gtag('config', googleAnalyticsId, {
-        user_id: uid,
-        user_properties: {
-          dept_code: _.map(user.departments || [], 'code'),
-          title: user.title,
-          uid
-        }
-      })
-    }
-    const track = (action, category, label?, id?) => {
-      // Report this event to Google Analytics if and only if certain conditions are true.
-      if (googleAnalyticsId && uid && !user.isAdmin) {
-        window.gtag('event', action, {
-          event_category: category,
-          event_label: label,
-          value: id
-        })
-      }
-    }
-    // BOA shortcuts
-    Vue.prototype.$ga = {
-      appointment: action => track(action, 'Appointment'),
-      cohort: (action, label?, id?) => track(action, 'Cohort', label, id),
-      course: (action, label?) => track(action, 'Course', label),
-      curated: (action, label?, id?) => track(action, 'Curated Group', label, id),
-      degreeProgress: (action, label?) => track(action, 'Degree Progress', label),
-      note: action => track(action, 'Advising Note'),
-      noteTemplate: action => track(action, 'Note Template'),
-      search: (action, label?) => track(action, 'Search', label),
-      student: (action, label?) => track(action, 'Student', label)
-    }
-    resolve()
-  })
 }
 
 export function isAdvisor(user) {
